fix: rm infura urls from lib
