+++ src/index.tsx
@@ -73,4 +73,4 @@ ReactDOM.render(
 if (process.env.REACT_APP_SERVICE_WORKER !== 'false') {
   serviceWorkerRegistration.register()
 }
-export { INFURA_NETWORK_URLS } from 'constants/chainInfo'
+export { INFURA_NETWORK_URLS } from 'constants/infura'
